requirements.txt
# Base dependencies
langchain>=0.1.0
langgraph>=0.0.10
openai>=1.5.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Web framework
fastapi>=0.100.0
uvicorn>=0.23.0

# Database
# sqlalchemy>=2.0.0
# pymongo>=4.0.0

# Testing
pytest>=7.0.0

# Development tools
black>=23.0.0
isort>=5.0.0
mypy>=1.0.0
flake8>=6.0.0